Gestura_GUI/requirements.txt
pywebview>=4.4
simpleaudio>=1.0.4
